requirements.txt
typer
langchain
langchain-anthropic
docker
pyyaml
python-dotenv
pytest
requests
duckduckgo-search
beautifulsoup4
lxml
google-generativeai
